Move `packed_simd.rs` to `vector` module
