+++ src/backend/vector/ifma/constants.rs
@@ -9,7 +9,7 @@
 
 //! This module contains constants used by the IFMA backend.
 
-use crate::backend::packed_simd::u64x4;
+use crate::backend::vector::packed_simd::u64x4;
 
 #[cfg(feature = "precomputed-tables")]
 use crate::window::NafLookupTable8;
